requirements.txt
APScheduler==3.7.0
gspread==3.6.0
python-dotenv==0.15.0
requests==2.25.1
psycopg2==2.8.6